refactor(lua/pay_kit): split protocols/ by protocol then scheme (Ludo review)
PR #141 review (lua/mpp/init.lua:22): the previous `schemes/` flat
namespace conflated protocol + scheme into one directory. Re-shape
to mirror the Ruby gem's per-protocol / per-scheme layout the
reviewer endorsed:

  resty/pay_kit/protocols/
  ├── mpp/init.lua                 # MPP wrapper (charge intent via mpp/server)
  └── x402/
      ├── init.lua                 # x402 adapter (offer + cosign + broadcast)
      └── exact/verify.lua         # 11-rule SVM-exact structural verifier

The split makes the protocol surface obvious at the directory level
and gives future schemes (`x402/upto`, `x402/batch`) and intents
(`mpp/session`, `mpp/subscription`) a place to live without
flattening into one folder.

Mechanical-only rename. All call sites updated:

- internal/dispatcher.lua now requires `resty.pay_kit.protocols.{mpp,x402}`
- harness/lua-server/server.lua reaches `protocols.mpp.set_splits_override`
- rockspec module map points at the new paths
- Tests + the broadcast spec's package.loaded eviction list use
  the new module names
- README repo-layout block reflects the new tree

No public-surface change; `resty.pay_kit.{configure,gate,...}`
stays the same. 528 specs pass, luacheck clean, lua server x10
matrix pairs green against surfpool.
