+++ lua/tests/pay_kit/x402_verify_negative_spec.lua
@@ -9,7 +9,7 @@ local base64 = require('mpp.util.base64_std')
 local base58 = require('mpp.util.base58')
 local tx_mod = require('mpp.methods.solana.transaction')
 local ata    = require('mpp.methods.solana.ata')
-local x402_verify = require('resty.pay_kit.schemes.x402_verify')
+local x402_verify = require('resty.pay_kit.protocols.x402.exact.verify')
 
 local function u64_le(n)
   local out = {}
